fixing multiple issues

+++ server/go/model_integration_fulfillment.go
@@ -16,14 +16,14 @@ type IntegrationFulfillment struct {
 	LineItemId string `json:"line_item_id"`
 
 	// Number of products fulfilled. The sum of quantities per `line_item_id` must match the total quantity of the line item.
-	Quantity float32 `json:"quantity"`
+	Quantity int32 `json:"quantity"`
 }
 
 // AssertIntegrationFulfillmentRequired checks if the required fields are not zero-ed
 func AssertIntegrationFulfillmentRequired(obj IntegrationFulfillment) error {
 	elements := map[string]interface{}{
 		"line_item_id": obj.LineItemId,
-		"quantity": obj.Quantity,
+		"quantity":     obj.Quantity,
 	}
 	for name, el := range elements {
 		if isZero := IsZeroValue(el); isZero {
